Accept changes which seem fine and likely just caused by evolving libraries

+++ Tests/CSharp/ExpressionTests/ExpressionTests.cs
@@ -319,7 +319,7 @@ public partial class VisualBasicClass
     }
 }
 1 target compilation errors:
-CS1503: Argument 1: cannot convert from 'object' to 'double'");
+CS1503: Argument 1: cannot convert from 'object' to 'decimal'");
     }
 
     [Fact]
